+++ DeepL/Model/WriteResult.cs
@@ -1,4 +1,4 @@
-// Copyright 2022 DeepL SE (https://www.deepl.com)
+// Copyright 2025 DeepL SE (https://www.deepl.com)
 // Use of this source code is governed by an MIT
 // license that can be found in the LICENSE file.
 
@@ -6,26 +6,24 @@ using System.Text.Json.Serialization;
 
 namespace DeepL.Model {
   /// <summary>The result of a text translation.</summary>
-  public sealed class TextResult {
-    /// <summary>Initializes a new instance of <see cref="TextResult" />.</summary>
-    /// <param name="text">Translated text.</param>
+  public sealed class WriteResult {
+    /// <summary>Initializes a new instance of <see cref="WriteResult" />.</summary>
+    /// <param name="text">improved text.</param>
     /// <param name="detectedSourceLanguageCode">The detected language code of the input text.</param>
-    /// <param name="billedCharacters">The number of characters billed for the text.</param>
-    /// <param name="modelTypeUsed">Indicates the translation model used, but is null unless the modelType option is specified.</param>
+    /// <param name="targetLanguageCode">The selected target language in the request.</param>
     /// <remarks>
     ///   The constructor for this class (and all other Model classes) should not be used by library users. Ideally it
     ///   would be marked <see langword="internal" />, but needs to be <see langword="public" /> for JSON deserialization.
     ///   In future this function may have backwards-incompatible changes.
     /// </remarks>
     [JsonConstructor]
-    public TextResult(string text, string detectedSourceLanguageCode, int billedCharacters, string? modelTypeUsed) {
+    public WriteResult(string text, string detectedSourceLanguageCode, string targetLanguageCode) {
       Text = text;
       DetectedSourceLanguageCode = LanguageCode.Standardize(detectedSourceLanguageCode);
-      BilledCharacters = billedCharacters;
-      ModelTypeUsed = modelTypeUsed;
+      TargetLanguageCode = targetLanguageCode;
     }
 
-    /// <summary>The translated text.</summary>
+    /// <summary>The improved text.</summary>
     public string Text { get; }
 
     /// <summary>The language code of the source text detected by DeepL.</summary>
@@ -33,15 +31,11 @@ namespace DeepL.Model {
     public string DetectedSourceLanguageCode { get; }
 
     /// <summary>The number of characters billed for the text.</summary>
-    [JsonPropertyName("billed_characters")]
-    public int BilledCharacters { get; }
+    [JsonPropertyName("target_language")]
+    public string TargetLanguageCode { get; }
 
-    /// <summary>Indicated translation model used, if available.</summary>
-    [JsonPropertyName("model_type_used")]
-    public string? ModelTypeUsed { get; }
-
-    /// <summary>Returns the translated text.</summary>
-    /// <returns>The translated text.</returns>
+    /// <summary>Returns the improved text.</summary>
+    /// <returns>The improved text.</returns>
     public override string ToString() => Text;
   }
 }