+++ DeepL/WritingTone.cs
@@ -1,0 +1,18 @@
+// Copyright 2025 DeepL SE (https://www.deepl.com)
+// Use of this source code is governed by an MIT
+// license that can be found in the LICENSE file.
+
+namespace DeepL {
+  /// <summary>Represents the tone the improved text should be in in a rephrase request.</summary>
+  public enum WritingTone {
+    Confident,
+    Default,
+    Diplomatic,
+    Enthusiastic,
+    Friendly,
+    PreferConfident,
+    PreferDiplomatic,
+    PreferEnthusiastic,
+    PreferFriendly,
+  }
+}